Add files via upload
Data Cleaning
Data cleaning is an important process for businesses because it ensures organisation, accuracy, productivity, cost savings, and improved data mapping. Every company believes that its data is clean, but this is not always the case. There are 5 steps to cleaning data:
●
Calculate 5 number summary: This involves calculating the minimum, first quartile, median, third quartile, and maximum values for each column of data. This can help to identify outliers and missing values.
●
Impute missing values: This involves filling in missing values with appropriate strategies, such as the mean, median, or constant values.
●
Correct or remove outliers: Outliers are extreme values that can affect model performance. These outliers can be corrected or removed.
●
Remove rows of high missing data: If a row of data has a lot of missing values, it may be best to remove it altogether.
●
Copy your data: It is always a good idea to make a copy of your data before you start cleaning it. This way, you will have a backup if you make any mistakes.